Extract indexed PNG encoder into IndexedPngWriter class
Main.java carried ~130 lines of low-level PNG chunk writing inlined
into its process(FileHandle, int) method: the SIGNATURE / IHDR / IDAT
/ IEND / PLTE / TRNS constants, five private instance fields for the
reusable Deflater and line buffers, the full write loop, and a
ChunkBuffer inner class. None of that is font-generation logic —
it's a specialised 8-bit indexed PNG encoder — and it made Main
harder to navigate by padding out the file with binary format code.

Move it all into a new IndexedPngWriter class in the same package.
The writer stays stateful across calls (line buffers and Deflater
are reused between atlases, matching the original behavior) so Main
holds a single reusable IndexedPngWriter instance. Main.process()
now handles only the fontwriter-specific pipeline steps — transparency
corner validation, the white-rectangle patch in the bottom-right, and
the MSDF fork — and delegates the actual write to the new class.

- Add IndexedPngWriter with write(FileHandle, Pixmap, int) and the
  private ChunkBuffer helper class
- Move PNG chunk constants, Deflater, and line buffer fields from
  Main into IndexedPngWriter (as private fields/constants)
- Replace the inlined write body in Main.process() with a single
  indexedPngWriter.write(file, pm, rgba) call
- Remove the ChunkBuffer inner class from Main
- Drop now-unused CRC32, CheckedOutputStream, Deflater, and
  DeflaterOutputStream imports from Main

Pure extraction — no behavioral changes. Output bytes are identical
to the previous version; the writer's state reuse semantics are
preserved.
