+++ core/src/main/java/com/github/tommyettinger/Main.java
@@ -21,29 +21,13 @@ import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.zip.CRC32;
-import java.util.zip.CheckedOutputStream;
-import java.util.zip.Deflater;
-import java.util.zip.DeflaterOutputStream;
 
 import static java.awt.Font.TRUETYPE_FONT;
 
 /** {@link com.badlogic.gdx.ApplicationListener} implementation shared by all platforms. */
 public class Main extends ApplicationAdapter {
 
-    static private final byte[] SIGNATURE = {(byte)137, 80, 78, 71, 13, 10, 26, 10};
-    static private final int IHDR = 0x49484452, IDAT = 0x49444154, IEND = 0x49454E44,
-            PLTE = 0x504C5445, TRNS = 0x74524E53;
-    static private final byte COLOR_INDEXED = 3;
-    static private final byte COMPRESSION_DEFLATE = 0;
-    static private final byte INTERLACE_NONE = 0;
-    static private final byte FILTER_NONE = 0;
-
-    private final ChunkBuffer buffer = new ChunkBuffer(65536);
-    private final Deflater deflater = new Deflater(0);
-    private ByteArray curLineBytes;
-    private ByteArray prevLineBytes;
-    private int lastLineLen;
+    private final IndexedPngWriter indexedPngWriter = new IndexedPngWriter();
     private FontwriterConfig config;
 
     private SpriteBatch batch;
@@ -780,115 +764,7 @@ public class Main extends ApplicationAdapter {
             return;
         }
 
-        OutputStream output = file.write(false);
-        try {
-            DeflaterOutputStream deflaterOutput = new DeflaterOutputStream(buffer, deflater);
-            DataOutputStream dataOutput = new DataOutputStream(output);
-            try {
-                dataOutput.write(SIGNATURE);
-
-                buffer.writeInt(IHDR);
-                buffer.writeInt(pm.getWidth());
-                buffer.writeInt(pm.getHeight());
-                buffer.writeByte(8); // 8 bits per component.
-                buffer.writeByte(COLOR_INDEXED);
-                buffer.writeByte(COMPRESSION_DEFLATE);
-                buffer.writeByte(FILTER_NONE);
-                buffer.writeByte(INTERLACE_NONE);
-                buffer.endChunk(dataOutput);
-
-                buffer.writeInt(PLTE);
-                for (int i = 0; i < 256; i++) {
-                    buffer.write(rgba >>> 24);
-                    buffer.write(rgba >>> 16 & 255);
-                    buffer.write(rgba >>>  8 & 255);
-                }
-                buffer.endChunk(dataOutput);
-
-                buffer.writeInt(TRNS);
-
-                for (int i = 0; i < 256; i++) {
-                    buffer.write(i);
-                }
-
-                buffer.endChunk(dataOutput);
-                buffer.writeInt(IDAT);
-                deflater.reset();
-
-                int lineLen = pm.getWidth();
-                byte[] curLine, prevLine;
-                if (curLineBytes == null) {
-                    curLine = (curLineBytes = new ByteArray(lineLen)).items;
-                    prevLine = (prevLineBytes = new ByteArray(lineLen)).items;
-                } else {
-                    curLine = curLineBytes.ensureCapacity(lineLen);
-                    prevLine = prevLineBytes.ensureCapacity(lineLen);
-                    for (int i = 0, n = lastLineLen; i < n; i++) {
-                        prevLine[i] = 0;
-                    }
-                }
-
-                lastLineLen = lineLen;
-
-                int color;
-                boolean noWarningNeeded = false;
-                for (int y = 0; y < h; y++) {
-                    for (int x = 0; x < w; x++) {
-                        color = pm.getPixel(x, y);
-                        // this block may need to be commented out if a font uses non-white grayscale colors.
-                        if(noWarningNeeded || ((color & 255) != 0 && (color & 0xFFFFFF00) != 0xFFFFFF00)) {
-                            System.out.println("PROBLEM WITH " + file);
-                            System.out.printf("Problem color: 0x%08X\n", color);
-                            System.out.println("Position: " + x + "," + y);
-                            noWarningNeeded = true;
-                        }
-                        curLine[x] = (byte) (color & 255);
-                    }
-
-                    deflaterOutput.write(FILTER_NONE);
-                    deflaterOutput.write(curLine, 0, lineLen);
-
-                    byte[] temp = curLine;
-                    curLine = prevLine;
-                    prevLine = temp;
-                }
-                deflaterOutput.finish();
-                buffer.endChunk(dataOutput);
-
-                buffer.writeInt(IEND);
-                buffer.endChunk(dataOutput);
-
-                output.flush();
-            } catch (IOException e) {
-                Gdx.app.error("transparency", e.getMessage());
-            }
-        } finally {
-            StreamUtils.closeQuietly(output);
-        }
-    }
-
-    static class ChunkBuffer extends DataOutputStream {
-        final ByteArrayOutputStream buffer;
-        final CRC32 crc;
-
-        ChunkBuffer(int initialSize) {
-            this(new ByteArrayOutputStream(initialSize), new CRC32());
-        }
-
-        private ChunkBuffer(ByteArrayOutputStream buffer, CRC32 crc) {
-            super(new CheckedOutputStream(buffer, crc));
-            this.buffer = buffer;
-            this.crc = crc;
-        }
-
-        public void endChunk(DataOutputStream target) throws IOException {
-            flush();
-            target.writeInt(buffer.size() - 4);
-            buffer.writeTo(target);
-            target.writeInt((int) crc.getValue());
-            buffer.reset();
-            crc.reset();
-        }
+        indexedPngWriter.write(file, pm, rgba);
     }
 
 }
